Slg dev win32 (#257)
* revising scan_info (again)

* more cleanup of options; distinct character count for sbuf; scan_aes now disables AES192 by default

* updated

* Addresses #234

* added hello world

* fixed conflict

* overflow bug in scan_aes

* expanded ETA

* fixed conflict

* fixed conflict

* removed POINT1

* updated readme

* added hello_win64.exe

* added starts_with

* properly records filename in feature recorder

* refactored main() out from bulk_extractor.cpp, so that we can do end-to-end testing on the executable

* created distinct_character_counter

* speed up sbuf a bit

* fixed carving extensions

* added support for scanner `carve` flag

* added end-to-end test of bulk_extractor_main()

* add installation of xmllint

* fixed handling of matcher

* finally all tests pass

* option processing improvements

* massive rewrite to use cxxopt instead of getopt

* further cleaned up option processing

* making article a subproject

* addressed issue with scan_hiberfile

* updated be13_api

* speed up AES finder.

* updated be13_api

* added initialization

* Reapplied #177 and #175

* whoops

* path_printer now has a test

* signed issue

* update

* handles disk full

* initial support for mingw

* updates for win32

* more for _WIN32

* added pcap_writer

* update

* updated for new dfxml

* fixed self-tests

* fixed image_process for MacOS. Need to test on the other platforms

+++ src/scan_aes.cpp
@@ -37,13 +37,13 @@
 
 /* old aes.h file */
 
-static const u_int AES128_KEY_SIZE  =               16; //  Size of a 128-bit AES key, in bytes
-static const u_int AES192_KEY_SIZE  =               24; // Size of a 192-bit AES key, in bytes
-static const u_int AES256_KEY_SIZE  =               32; // Size of a 256-bit AES key, in bytes
+static const size_t AES128_KEY_SIZE  =               16; //  Size of a 128-bit AES key, in bytes
+static const size_t AES192_KEY_SIZE  =               24; // Size of a 192-bit AES key, in bytes
+static const size_t AES256_KEY_SIZE  =               32; // Size of a 256-bit AES key, in bytes
 
-static const u_int AES128_KEY_SCHEDULE_SIZE =      176; // Size of a 128-bit AES key schedule, in bytes
-static const u_int AES192_KEY_SCHEDULE_SIZE =      208; // Size of a 128-bit AES key schedule, in bytes
-static const u_int AES256_KEY_SCHEDULE_SIZE =      240; // Size of a 128-bit AES key schedule, in bytes
+static const size_t AES128_KEY_SCHEDULE_SIZE =      176; // Size of a 128-bit AES key schedule, in bytes
+static const size_t AES192_KEY_SCHEDULE_SIZE =      208; // Size of a 128-bit AES key schedule, in bytes
+static const size_t AES256_KEY_SCHEDULE_SIZE =      240; // Size of a 128-bit AES key schedule, in bytes
 
 
 // Determines whether or not data represents valid
